write to correct directory

+++ src/TwitterFileService.java
@@ -8,37 +8,44 @@ public class TwitterFileService {
     public void writeUser(User user, Configuration configuration) throws IOException {
 
         // needs to be modified to use correct directory.
-        File file = new File(configuration.getUserFile());
+        File file = new File(configuration.getDataDirectory(), configuration.getUserFile());
 
         // only creates a new file if it doesn't exist
         file.createNewFile();
-        boolean found = false;
-        try (BufferedReader br = new BufferedReader(new FileReader(file))) {
-            String line;
-            while ((line = br.readLine()) != null)  // classic way of reading a file line-by-line
-                if (line.equals("@" + user.getScreenName() + "\t"
-                        + user.getLocation() + "\t"
-                        + user.getDescription().replaceAll("\n", " ") + "\t"
-                        + user.getFollowersCount())) {
 
-                    found = true;
+        if (file.exists() || file.getParentFile().mkdirs()) {
 
-                    break;  // if the text is present, we do not have to read the rest after all
-                }
-        } catch (FileNotFoundException fnfe) {
-            fnfe.printStackTrace();
-        }
+            boolean found = false;
+            try (BufferedReader br = new BufferedReader(new FileReader(file))) {
+                String line;
+                while ((line = br.readLine()) != null)  // classic way of reading a file line-by-line
+                    if (line.equals("@" + user.getScreenName() + "\t"
+                            + user.getLocation() + "\t"
+                            + user.getDescription().replaceAll("\n", " ") + "\t"
+                            + user.getFollowersCount())) {
 
-        if (!found) {
-            try (PrintWriter pw = new PrintWriter(new FileWriter(file, true))) {
-                pw.println("@" + user.getScreenName() + "\t"
-                        + user.getLocation() + "\t"
-                        + user.getDescription().replaceAll("\n", " ") + "\t"
-                        + user.getFollowersCount());
-                pw.flush();
-            } catch (IOException e) {
-                e.printStackTrace();
+                        found = true;
+
+                        break;  // if the text is present, we do not have to read the rest after all
+                    }
+
+            } catch (FileNotFoundException fnfe) {
+                fnfe.printStackTrace();
             }
+
+            if (!found) {
+                try (PrintWriter pw = new PrintWriter(new FileWriter(file, true))) {
+                    pw.println("@" + user.getScreenName() + "\t"
+                            + user.getLocation() + "\t"
+                            + user.getDescription().replaceAll("\n", " ") + "\t"
+                            + user.getFollowersCount());
+                    pw.flush();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        } else {
+            throw new RuntimeException("Couldn't create / get file in the directory");
         }
     }
 
@@ -65,29 +72,33 @@ public class TwitterFileService {
 
     public void writeTweet(Status tweet, boolean retweet, Configuration configuration) throws IOException {
 
-        // needs to be modified to use correct directory.
-        File file = new File(configuration.getDataFile());
+        File file = new File(configuration.getDataDirectory(), configuration.getDataFile());
 
-        // only creates a new file if it doesn't exist
-        file.createNewFile();
+        if (file.exists() || file.getParentFile().mkdirs()) {
 
-        if (!isTweetInFile(tweet, file.getAbsolutePath())) {
-            try (PrintWriter pw = new PrintWriter(new FileWriter(file, true))) {
-                String tweetText;
-                if (retweet)
-                    tweetText = tweet.getRetweetedStatus().getText();
-                else
-                    tweetText = tweet.getText();
+            // only creates a new file if it doesn't exist
+            file.createNewFile();
 
-                pw.println(tweet.getId() + "\t"
-                        + "@" + tweet.getUser().getScreenName() + "\t"
-                        + tweetText.replaceAll("\n", " ") + "\t"
-                        + tweet.getRetweetCount() + "\t"
-                        + tweet.getCreatedAt());
-                pw.flush();
-            } catch (IOException e) {
-                e.printStackTrace();
+            if (!isTweetInFile(tweet, file.getAbsolutePath())) {
+                try (PrintWriter pw = new PrintWriter(new FileWriter(file, true))) {
+                    String tweetText;
+                    if (retweet)
+                        tweetText = tweet.getRetweetedStatus().getText();
+                    else
+                        tweetText = tweet.getText();
+
+                    pw.println(tweet.getId() + "\t"
+                            + "@" + tweet.getUser().getScreenName() + "\t"
+                            + tweetText.replaceAll("\n", " ") + "\t"
+                            + tweet.getRetweetCount() + "\t"
+                            + tweet.getCreatedAt());
+                    pw.flush();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
+        } else {
+            throw new RuntimeException("Couldn't create / get file in the directory");
         }
     }
